!refactor: mayor nvim refactor
Completely migrated from init.vim to init.lua. I'm quite certain there
are some improvements to be made in this, but for now it works.

TODO: Still need to refactor some *.vim files. Also, do all the files in
lua/robin need to be there? Can't they be moved to the after/plugin
directory and be auto sourced?